requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
defusedxml==0.7.1
distlib==0.3.3
Django==4.0.2
django-ckeditor==6.2.0
django-cors-headers==3.10.0
django-crum==0.7.9
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.12.4
filelock==3.4.0
hashids==1.3.1
httpie==3.0.2
idna==3.3
multidict==6.0.2
Pillow==9.0.1
platformdirs==2.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
Pygments==2.11.2
PySocks==1.7.1
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.10.0
